+++ react-hook/index.cjs
@@ -1,0 +1,3 @@
+'use strict'
+
+exports.useInput = require('../commonjs/react/useInput.js').default