+++ source/templateFormatter.js
@@ -9,28 +9,24 @@ import close_braces from './closeBraces.js'
 // If the `template` can only be partially filled
 // then it is cut off.
 //
-// If `should_close_braces` is `true`,
+// If `shouldCloseBraces` is `true`,
 // then it will also make sure all dangling braces are closed,
 // e.g. "8 (8" -> "8 (8  )" (iPhone style phone number input).
 //
-export default function(template, placeholder = 'x', should_close_braces)
-{
-	if (!template)
-	{
+export default function(template, placeholder = 'x', shouldCloseBraces) {
+	if (!template) {
 		return value => ({ text: value })
 	}
 
-	const characters_in_template = count_occurences(placeholder, template)
+	const placeholdersCountInTemplate = count_occurences(placeholder, template)
 
-	return function(value)
-	{
-		if (!value)
-		{
+	return function(value) {
+		if (!value) {
 			return { text: '', template }
 		}
 
-		let value_character_index = 0
-		let filled_in_template = ''
+		let characterIndexInValue = 0
+		let templateWithFilledInPlaceholders = ''
 
 		// Using `.split('')` here instead of normal `for ... of`
 		// because the importing application doesn't neccessarily include an ES6 polyfill.
@@ -38,38 +34,33 @@ export default function(template, placeholder = 'x', should_close_braces)
 		// (the ones consisting of four bytes)
 		// but template placeholder characters don't fall into that range
 		// and appending UTF-8 characters to a string in parts still works.
-		for (const character of template.split(''))
-		{
-			if (character !== placeholder)
-			{
-				filled_in_template += character
+		for (const character of template.split('')) {
+			if (character !== placeholder) {
+				templateWithFilledInPlaceholders += character
 				continue
 			}
 
-			filled_in_template += value[value_character_index]
-			value_character_index++
+			templateWithFilledInPlaceholders += value[characterIndexInValue]
+			characterIndexInValue++
 
 			// If the last available value character has been filled in,
 			// then return the filled in template
 			// (either trim the right part or retain it,
 			//  if no more character placeholders in there)
-			if (value_character_index === value.length)
-			{
+			if (characterIndexInValue === value.length) {
 				// If there are more character placeholders
 				// in the right part of the template
 				// then simply trim it.
-				if (value.length < characters_in_template)
-				{
+				if (value.length < placeholdersCountInTemplate) {
 					break
 				}
 			}
 		}
 
-		if (should_close_braces)
-		{
-			filled_in_template = close_braces(filled_in_template, template)
+		if (shouldCloseBraces) {
+			templateWithFilledInPlaceholders = close_braces(templateWithFilledInPlaceholders, template)
 		}
 
-		return { text: filled_in_template, template }
+		return { text: templateWithFilledInPlaceholders, template }
 	}
 }