+++ source/templateParser.js
@@ -1,19 +1,16 @@
 import { count_occurences } from './helpers.js'
 
-export default function(template, placeholder, parse)
-{
-	if (typeof placeholder === 'function')
-	{
+// `placeholder` argument is optional. By default, it's "x".
+export default function(template, placeholder, parse) {
+	if (typeof placeholder === 'function') {
 		parse = placeholder
 		placeholder = 'x'
 	}
 
-	const max_characters = count_occurences(placeholder, template)
+	const placeholdersCountInTemplate = count_occurences(placeholder, template)
 
-	return function(character, value)
-	{
-		if (value.length < max_characters)
-		{
+	return function(character, value) {
+		if (value.length < placeholdersCountInTemplate) {
 			return parse(character, value)
 		}
 	}